Moved env to app folder & removed EAS_SLUG
