+++ apps/expo/.env.example
@@ -1,4 +1,3 @@
 EAS_OWNER="timothymiller"
 EAS_PROJECT_ID="85fc6ccd-0ce1-4e4d-804c-b15df989f97e"
-EAS_SLUG="t4-app"
 EXPO_NO_TELEMETRY=true
