+++ package.json
@@ -43,7 +43,6 @@
     "@t3-oss/env-core": "^0.7.1",
     "check-dependency-version-consistency": "^4.1.0",
     "cross-env": "^7.0.3",
-    "eas-cli": "^5.6.0",
     "eslint": "^8.52.0",
     "next-superjson-plugin": "^0.5.9",
     "npm-run-all": "^4.1.5",
