+++ packages/app/features/sign-up/screen.native.tsx
@@ -9,7 +9,7 @@ import * as WebBrowser from 'expo-web-browser'
 import { getInitialURL } from 'expo-linking'
 import { Platform } from 'react-native'
 import { initiateAppleSignIn } from 'app/utils/supabase/appleAuth'
-import { env } from "../../../env.mjs";
+import { env } from "app/env";
 
 export const SignUpScreen = (): React.ReactNode => {
   const { replace } = useRouter()
